apps/flask_apps/samgov_search_flask_app/requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
email_validator
